primitive generator renamed and updated
